INFO 201 a2 R code and written responses
Edits made to the homework 2 files to address questions posed regarding the CountLove data set that describes statistical values corresponding to protests in the United States. The R code works to answer the posed questions and the readme addresses the written responses to some of the questions that the code was able to answer.